pyproject.toml
[tool.poetry]
name = "hacklifebalance"
version = "0.1.0"
description = ""
authors = ["klim <klimrus61@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
black = "^23.10.0"
isort = "^5.12.0"
django = "^4.2.6"
djangorestframework = "^3.14.0"
environ = "^1.0"
rest-framework-simplejwt = "^0.0.2"
channels = "^4.0.0"
django-environ = "^0.11.2"
drf-spectacular = "^0.26.5"
djangorestframework-simplejwt = "^5.3.0"
requests = "^2.31.0"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
